+++ CHANGELOG.md
@@ -5,6 +5,17 @@ file is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and
 this project adheres to [Semantic
 Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- `RandomizedPCA`, a randomized, truncated PCA.
+
+### Fixed
+
+- `Pca::explained_variance_ratio` returns the ratios for the principal
+  components only.
+
 ## [0.1.1] - 2020-05-21
 
 ### Added
@@ -17,5 +28,6 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 - Principal component analysis (PCA).
 
+[Unreleased]: https://github.com/petabi/petal-decomposition/compare/0.1.1...master
 [0.1.1]: https://github.com/petabi/petal-decomposition/compare/0.1.0...0.1.1
 [0.1.0]: https://github.com/petabi/petal-decomposition/tree/0.1.0